Move registartionUser and loginUser resolvers to authentications.js. Add validator for checking test environment. Add -- --runInBand flag to npm test.
