+++ server/config/config.json
@@ -14,8 +14,6 @@
     }
   },
   "test": {
-    "username": "birhanu",
-    "password": null,
     "database": "toolsio_test",
     "host": "127.0.0.1",
     "dialect": "postgres",
