+++ server/src/resolvers/user.js
@@ -40,147 +40,6 @@ export default {
   },
 
   Mutation: {
-    loginUser: (parent, { email, password }, { models, subdomain, SECRET, SECRET2 }) => 
-      loginUserWithToken(email, password, models, subdomain, SECRET, SECRET2),
-    
-    registerUser: async (parent, args, { models }) => {
-
-      const { firstName, lastName, email, password } = args
-      const { subdomain, industry } = args
-
-      try {
-        const account = await models.Account.findOne( { where: { subdomain } }, { raw: true })
-
-        if (account) {
-          return {
-            success: false,
-            errors: [
-              {
-                path: 'subdomain',
-                message: 'Subdomain is already taken'
-              }
-            ]
-          }
-        } else {
-          const response = await models.sequelize.transaction(async (transaction) => {
-            
-            // Creates a schema
-            await models.sequelize.createSchema(subdomain)
-
-            // Migrate users model 
-            await models.User.sync({schema: subdomain})
-            
-            // Sync the reset schemas asynchronously
-            Promise.all(
-              map(Object.keys(models), (key) => {
-                if (['sequelize', 'Sequelize', 'Account', 'User'].includes(key)) return
-                models[key].sync({schema: subdomain})
-              })
-            )
-
-            const user = await  models.User.create({ firstName, lastName, email, password }, { searchPath: subdomain, transaction })
-            const account = await models.Account.create({ subdomain, industry, owner: user.id }, { transaction })
-
-            return { user, account } 
-          })
-
-          // Create emailToken
-          jwt.sign({
-            id: response.user.dataValues.id,
-            email: response.user.dataValues.email
-          }, jwtConfig.jwtSecret1, { expiresIn: '60d' }, (err, emailToken) => {
-            if (err) {
-              console.log('err token: ', err)
-            }
-            
-            const url = `http://${response.account.subdomain}.lvh.me:3000/login/confirmation/${emailToken}`
-
-            transporter.sendMail({
-              to: userCreated.get('email'),
-              subject: 'Confirm Email (Toolsio)',
-              html: `Please click this link to confirm your email: <a href="${url}">${url}</a>`
-            })
-          })
-      
-          return {
-            success: true,
-            account: response.account
-          }
-        }
-      } catch(err) {
-        console.log('err: ', err)
-        return {
-          success: false,
-          errors: formatErrors(err, models)
-        }       
-      }
-
-    },
-
-    registerInvitedUser: async (parent, args, { models }) => {
-
-      const { firstName, lastName, email, password, token } = args
-
-      const { account } = jwt.verify(token, jwtConfig.jwtSecret1)
-
-      try {
-        const accountLocal = await models.Account.findOne( { where: { subdomain: account } }, { raw: true })
-
-        if (accountLocal) {
-
-          try {
-            const user = await  models.User.schema(accountLocal.subdomain).create({ firstName, lastName, email, password }, { searchPath: accountLocal.subdomain })
-
-            // Create emailToken
-            jwt.sign({
-              id: user.dataValues.id,
-              email: user.dataValues.email
-            }, jwtConfig.jwtSecret1, { expiresIn: '60d' }, (err, emailToken) => {
-              if (err) {
-                console.log('err token: ', err)
-              }
-              
-              const url = `http://${accountLocal.dataValues.subdomain}.lvh.me:3000/login/confirmation/${emailToken}`
-
-              transporter.sendMail({
-                to: userCreated.get('email'),
-                subject: 'Confirm Email (Toolsio)',
-                html: `Please click this link to confirm your email: <a href="${url}">${url}</a>`
-              })
-            })
-        
-            return {
-              success: true,
-              account: accountLocal
-            }
-          } catch(err) {
-            console.log('err: ', err)
-            return {
-              success: false,
-              errors: formatErrors(err, models)
-            }       
-          }
-        } else {
-          return {
-            success: false,
-            errors: [
-              {
-                path: 'subdomain',
-                message: `Account *${account}* to which you are invited doesn't exist. You can create it as your own new account by going to sign up page`
-              }
-            ]
-          }
-        }
-      } catch(err) {
-        console.log('err: ', err)
-        return {
-          success: false,
-          errors: formatErrors(err, models)
-        }       
-      }
-
-    },
-
     sendInvitation: requiresAuth.createResolver((parent, args, { subdomain }) => {
       
       let emailToken
