+++ server/src/types/user.js
@@ -20,22 +20,9 @@ export default `
     isConfirmed: Boolean!
   }
 
-  type RegisterResponse {
-    success: Boolean!
-    account: Account
-    errors: [Error!]
-  }
-
   type UpdateResponse {
     success: Boolean!
     user: User
-    errors: [Error!]
-  }
-
-  type LoginResponse {
-    success: Boolean!
-    authToken: String
-    refreshAuthToken: String
     errors: [Error!]
   }
 
@@ -60,17 +47,10 @@ export default `
   }
 
   type Mutation {
-    registerUser(firstName: String, lastName: String, email: String!, password: String!, 
-      subdomain: String!, industry: String!): RegisterResponse!
-
     updateUser(firstName: String, lastName: String, email: String!, password: String, 
       avatarUrl: String): UpdateResponse!
 
-    loginUser(email: String!, password: String!): LoginResponse!
-
     sendInvitation(email: String!): SendInvitationResponse!
-
-    registerInvitedUser(firstName: String, lastName: String, email: String!, password: String!, token: String!): RegisterResponse!
 
     s3SignAvatar(fileName: String!, fileType: String!): S3SignAvatarResponse!
   }
